feat(reviewtransaction): delete the ReconcileInvalidRecoveryEdge provider
WU8 (S3b), completing the reconcile-authority retirement started in WU7.

## What retires

ReconcileInvalidRecoveryEdge (compact_reconcile.go) and its four
tightly-coupled helpers -- compactReconcileAuthorizationSchema,
CompactReconcileRequest, compactReconcileAuthorizationBinding,
compactCombinedReconcileAuthorizationBinding, compactNonReconcilableContinuation
-- confirmed to have zero non-test callers by WU7's own deadcode ratchet
uptick (244 -> 248), which named exactly these symbols. This slice deletes
them and, as predicted, the ratchet reverses that uptick exactly (248 ->
244, net -4).

## Retained (compact_reconcile.go survives as a smaller, non-reconcile file)

classifyCompactRecoveryEdgeAnomalies (3 live callers: authority_disposition_plan.go,
compact_inspect.go, compact_batch_reconcile_plan.go), compactAbandonCommandText/
compactAbandonBlockerText (compact_reclaim.go, compact_inspect.go),
CompactInvalidRecoveryEdgeProof/CompactMalformedRecoveryAuthorizationProof
(compact_reclaim.go's CompactReclaimRecord still carries these as optional
historical fields for records already on disk), and the anomaly-class
constants/classification type the classifier owns. File keeps its name --
renaming risks more than it clarifies given same-package implicit reuse --
but its header comment now says plainly what's left and why.

## Test file surgery (measured before splitting; fit in one slice)

compact_reconcile_test.go (1192L) deleted entirely EXCEPT
TestClassifyCompactRecoveryEdgeAnomalies (tests retained, live code) and
four fixture helpers reused by OTHER test files (confirmed by grep before
deleting, the same cross-file-test-infrastructure-reuse check WU4 and WU7
both required): writeCompactFixtureRecord (8 other files),
poisonedRecoveryFixture (compact_batch_reconcile_journal_test.go),
preContractRecoveryFixture + preContractFixtureAuthorization
(compact_damaged_store_exit_test.go, compact_batch_reconcile_plan_test.go,
compact_batch_reconcile_lock_test.go). All five moved to a new
compact_fixture_test.go rather than dying with their original home.

Four other test files had direct ReconcileInvalidRecoveryEdge calls, each
handled per its own actual role (never assumed):
- compact_damaged_store_exit_test.go: TestReconcileRefusesUnreadableSuccessorWithDiagnosis
  deleted outright -- its own shape (truncated record) is already covered,
  end to end, by a sibling subtest driving the retained `review reclaim`
  refusal instead.
- compact_forged_authorization_test.go: two tests partially rewritten --
  each already independently proved the SanctionedCompactRecoveryExits/
  read-only-inspection half of its story; only the redundant
  ReconcileInvalidRecoveryEdge confirmation step is removed. forgedReconcileRequest
  helper deleted (zero remaining callers).
- compact_abandon_test.go, compact_legacy_quarantine_test.go: each had ONE
  ReconcileInvalidRecoveryEdge call as one of several surfaces proving "no
  sanctioned exit" or "routes nowhere" -- removed, the remaining surfaces
  (reclaim, abandon, invalidate) still prove the identical conclusion.

## Ratchets, both directions confirmed by re-running --update

Deadcode: 248 -> 244 (net -4), exactly reversing WU7's own uptick as the
coordinator predicted -- "no new unreachable functions" on this run.
Refusal: baseline shrank by 9 entries (ReconcileInvalidRecoveryEdge's own
refusal sites, all now gone with the function).

## Test evidence

go build/vet/gofmt: clean. go test ./internal/reviewtransaction/... and
./internal/cli/... and the full root ./...: green except the one designed
TestLegacyReadOnlyGuardMutationVerbsUnreachable (RG.1b, unaffected -- WU8
touches no dispatch case). bench --axis all: 83 completed/1
unsupported(j57, pre-existing)/0 failed on a fresh binary; diffed
field-by-field against a binary built from this branch's own parent
(198cb33e) -- zero differing dimensions across all 85 journey rows.

Net line count: +333/-1600 (net -1267). No split needed -- the
reconcile-only subset measured well under the 1000-line budget once the
five shared fixtures were identified and excluded.

## Rollback boundary

Single commit; `git revert` restores the provider, its four helpers, and
every test file's reconcile-specific portion atomically.
