+++ .deadcode-baseline.txt
@@ -64,10 +64,6 @@ internal/reviewtransaction/candidate_decline.go	parseCandidateDeclineAuthorizati
 internal/reviewtransaction/candidate_relation.go	shadowPushBoundaryUnresolvable
 internal/reviewtransaction/candidate_relation.go	shadowRelationHasNoLiveCounterpart
 internal/reviewtransaction/compact_batch_reconcile_guard.go	validateCompactBatchReconcileRequest
-internal/reviewtransaction/compact_reconcile.go	ReconcileInvalidRecoveryEdge
-internal/reviewtransaction/compact_reconcile.go	compactCombinedReconcileAuthorizationBinding
-internal/reviewtransaction/compact_reconcile.go	compactNonReconcilableContinuation
-internal/reviewtransaction/compact_reconcile.go	compactReconcileAuthorizationBinding
 internal/reviewtransaction/compact_result_disposition.go	CompactResultDispositionAuthorization
 internal/reviewtransaction/gate.go	hashArtifactSource
 internal/reviewtransaction/gate.go	hashLedgerArtifactSource
