+++ .refusal-ratchet-baseline.txt
@@ -704,15 +704,6 @@ internal/reviewtransaction/compact_reclaim.go	"review quarantine refused unsafe 
 internal/reviewtransaction/compact_reclaim.go	"review reclaim requires a non-empty reason and actor"
 internal/reviewtransaction/compact_reconcile.go	"combined recovery anomalies do not re-derive exactly: %v"
 internal/reviewtransaction/compact_reconcile.go	"recovery edge fails outside the unchanged-target class and the pre-contract authorization class: %v"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority refused: %v.%s"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority refused: predecessor %q has another successor %q"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority refused: recovery edge for %q validates; the successor remains authoritative"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority refused: successor %q has its own successor %q"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority refused: successor %q is not a recovery successor"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority refused: successor %q names predecessor %q"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority requires a non-empty reason and actor"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority requires an exact maintainer authorization binding (schema %s over predecessor %s@%s and successor %s@%s)"
-internal/reviewtransaction/compact_reconcile.go	"review reconcile-authority requires distinct predecessor and successor lineages"
 internal/reviewtransaction/compact_reconcile.go	"successor %q records a %s binding bound to different content; that is corruption, not a pre-contract authorization"
 internal/reviewtransaction/compact_reconcile.go	"the pre-contract authorization is not the sole edge anomaly: %v"
 internal/reviewtransaction/compact_reconcile.go	"unchanged target is not the sole anomaly; successor %q records a %s binding bound to different content, which is corruption rather than a pre-contract authorization"
